+++ commhandler/src/sgr_commhandler/utils/jmespath_mapping.py
@@ -1,6 +1,5 @@
 import collections
 import json
-import logging
 import re
 import copy
 from typing import Any, Optional