+++ commhandler/CHANGELOG.md
@@ -17,10 +17,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - implemented REST interface basic authentication
 - implemented dynamic parameters in REST calls
 - scaling of Modbus data point values
+- implemented MQTT messaging interface
+- implemented JMESpath mappings in response queries and message filters
+- implemented regex in response queries and message filters
 
 ### Changed
 
 - requires updated SGr specification `2.1.20250902`
+- updated dependencies
+- modified `subscribe()` method to provide reference to data point
 
 
 ## [0.3.1] - 2025-04-16